vtt-sync: phase 5 extract normalizers from store.js into state/normalize/
store.js had grown to 2,121 lines, mixing the public store API
(createStore/getState/updateState/subscribe) with ~40 normalisation
helpers for every state shape. Split those helpers into focused
submodules under state/normalize/:

  helpers.js           shared primitive coercions
  pings.js             map ping normalization
  grid.js              grid/scene grid application
  templates.js         circle/rectangle/wall templates
  drawings.js          freehand drawings
  fog.js               fog-of-war normalization + snapshot helpers
  monsters.js          monster snapshot + player-view stripping
  combat.js            combat state / turn lock / groups
  scenes.js            scene folders and items
  tokens.js            token library + player-view restriction
  placements.js        placement entries + conditions + HP
  overlay.js           overlay layers + aggregate mask + sync
  scene-board-state.js aggregator for per-scene board state
  index.js             barrel re-export

store.js shrinks from 2,121 to ~150 lines and now only contains the
store API: state, initializeState, getState, subscribe, updateState,
notify. Public exports are unchanged (re-exported through store.js
for backward compatibility with existing importers).

Behavior unchanged. All 76 state tests pass.
