+++ dnd/vtt/assets/js/state/normalize/helpers.js
@@ -1,0 +1,127 @@
+export function toInt(value, fallback = 0) {
+  const parsed = Number.parseInt(value, 10);
+  if (Number.isFinite(parsed)) {
+    return parsed;
+  }
+
+  const numeric = Number(value);
+  if (Number.isFinite(numeric)) {
+    return Math.trunc(numeric);
+  }
+
+  return Math.trunc(fallback);
+}
+
+export function toNonNegativeInt(value, fallback = 0) {
+  if (typeof value === 'number' && Number.isFinite(value)) {
+    return Math.max(0, Math.trunc(value));
+  }
+
+  const parsed = Number.parseInt(value, 10);
+  if (Number.isFinite(parsed)) {
+    return Math.max(0, parsed);
+  }
+
+  return Math.max(0, Math.trunc(fallback));
+}
+
+export function toBoolean(value, fallback = false) {
+  if (typeof value === 'boolean') {
+    return value;
+  }
+
+  if (typeof value === 'number') {
+    if (!Number.isFinite(value)) {
+      return fallback;
+    }
+    return value !== 0;
+  }
+
+  if (typeof value === 'string') {
+    const normalized = value.trim().toLowerCase();
+    if (!normalized) {
+      return fallback;
+    }
+    if (normalized === 'true' || normalized === '1' || normalized === 'yes' || normalized === 'on') {
+      return true;
+    }
+    if (normalized === 'false' || normalized === '0' || normalized === 'no' || normalized === 'off') {
+      return false;
+    }
+    return fallback;
+  }
+
+  if (typeof value === 'object' && value !== null) {
+    return toBoolean(value.valueOf(), fallback);
+  }
+
+  return fallback;
+}
+
+export function toOptionalNumber(value) {
+  if (typeof value === 'number' && Number.isFinite(value)) {
+    return Math.trunc(value);
+  }
+  if (typeof value === 'string') {
+    const trimmed = value.trim();
+    if (!trimmed) {
+      return null;
+    }
+    const parsed = Number.parseInt(trimmed, 10);
+    if (Number.isFinite(parsed)) {
+      return parsed;
+    }
+    const numeric = Number(trimmed);
+    if (Number.isFinite(numeric)) {
+      return Math.trunc(numeric);
+    }
+  }
+  return null;
+}
+
+export function clampToFinite(value, fallback = 0, precision = 4) {
+  const parsed = Number(value);
+  if (!Number.isFinite(parsed)) {
+    return fallback;
+  }
+  if (!Number.isFinite(precision) || precision <= 0) {
+    return parsed;
+  }
+  const factor = 10 ** precision;
+  return Math.round(parsed * factor) / factor;
+}
+
+export function roundToPrecision(value, precision = 4) {
+  if (!Number.isFinite(value)) {
+    return 0;
+  }
+
+  const places = Number.isFinite(precision) ? Math.max(0, Math.trunc(precision)) : 0;
+  const factor = 10 ** places;
+  return Math.round(value * factor) / factor;
+}
+
+export function uniqueStringList(raw = []) {
+  if (!Array.isArray(raw)) {
+    return [];
+  }
+
+  const seen = new Set();
+  const result = [];
+
+  raw.forEach((value) => {
+    if (typeof value !== 'string') {
+      return;
+    }
+
+    const trimmed = value.trim();
+    if (!trimmed || seen.has(trimmed)) {
+      return;
+    }
+
+    seen.add(trimmed);
+    result.push(trimmed);
+  });
+
+  return result;
+}
