+++ dnd/vtt/assets/js/state/normalize/scenes.js
@@ -1,0 +1,17 @@
+export function normalizeScenes(raw = {}) {
+  if (Array.isArray(raw)) {
+    return { folders: [], items: raw };
+  }
+
+  const folders = Array.isArray(raw?.folders) ? raw.folders : [];
+  const items = Array.isArray(raw?.items)
+    ? raw.items
+    : Array.isArray(raw?.scenes)
+    ? raw.scenes
+    : [];
+
+  return {
+    folders: folders.filter((folder) => folder && typeof folder.id === 'string'),
+    items: items.filter((scene) => scene && typeof scene.id === 'string'),
+  };
+}
